+++ include/wx/wxsqlite3.h
@@ -3,7 +3,7 @@
 ** Purpose:     wxWidgets wrapper around the SQLite3 embedded database library.
 ** Author:      Ulrich Telle
 ** Created:     2005-07-14
-** Copyright:   (c) 2005-2025 Ulrich Telle
+** Copyright:   (c) 2005-2026 Ulrich Telle
 ** License:     LGPL-3.0+ WITH WxWindows-exception-3.1
 */